Cleaned code

+++ src/btg_file.cpp
@@ -102,8 +102,8 @@ std::vector<Triangles> BtgFile::generateTriangles()
         triangle.material = std::string("../resources/textures/btg/") + obj.first + std::string(".png");
         for (size_t i = 0; i < obj.second.size(); i++)
         {
-            if(obj.second[i].vertexIndex>=mVertices.getSize() ||
-               obj.second[i].textureCoordIndex>=mTextureCoordinates.getSize()) {
+            if(static_cast<size_t>(obj.second[i].vertexIndex)>=mVertices.getSize() ||
+               static_cast<size_t>(obj.second[i].textureCoordIndex)>=mTextureCoordinates.getSize()) {
                 std::cout << "Damage btg file" << std::endl;
                 return triangles;
             }
